move and update ProgramCompletionBanner
